Create putinwave_1.5.c
In this updated code, the get_file_size function is introduced to calculate the size of the file to hide or extract. It uses the fseek and ftell functions to determine the size by seeking to the end of the file and then returning the current position indicator. The calculated size is then used instead of the user input.

Note: It's important to ensure that the files being passed as arguments exist and are accessible before calling the get_file_size function.